Hide app secrets such as API keys in code.
Move secret keys out of the repository and hide them in the binary.

C-string -> String decoding for unsigned strings and buffers to
facilitate conversion of various C API strings into Swift.

Some cleanup in base64 coding utility.

A utility for converting hex strings into byte buffers.

Improved checks during AES routine and fix PKCS padding bug if
plain text fits block size.

+++ platform-independent/c/core/src/base64.h
@@ -69,7 +69,7 @@ size_t mpw_base64_decode_max(const char *b64Text);
   * @param plainBuf a byte buffer, size should be at least mpw_base64_decode_max(b64Text)
   * @return The amount of bytes that were written to plainBuf.
   */
-int mpw_base64_decode(uint8_t *plainBuf, const char *b64Text);
+size_t mpw_base64_decode(uint8_t *plainBuf, const char *b64Text);
 
 /**
  * @return The amount of characters needed to encode a plainBuf of the given size as base-64 (excluding the terminating NUL).
@@ -79,4 +79,4 @@ size_t mpw_base64_encode_max(size_t plainSize);
   * @param b64Text a character buffer, size should be at least mpw_base64_encode_max(plainSize) + 1
   * @return The amount of characters that were written to b64Text, excluding the terminating NUL.
   */
-int mpw_base64_encode(char *b64Text, const uint8_t *plainBuf, size_t plainSize);
+size_t mpw_base64_encode(char *b64Text, const uint8_t *plainBuf, size_t plainSize);
